added health bar for player and collision damage

+++ Assets/Scripts/HealthBar.cs
@@ -6,12 +6,12 @@ using UnityEngine.UI;
 public class HealthBar : MonoBehaviour
 {
     public Slider slider;
-    public void setMaxHealth(int health){
+    public void setMaxHealth(float health){
         slider.maxValue = health;
         slider.value = health;
         
     }
-    public void setHealth(int health){
+    public void setHealth(float health){
         slider.value = health;
     }
 }
